requirements.txt
numpy
scipy
pillow
natsort
opencv-python
matplotlib
pandas
openpyxl
pyarrow #to save df as feather
scikit-image
scikit-learn
jupyterlab
openslide-python
docopt
future
imgaug
psutil
termcolor
tqdm
tensorboard
tensorboardx
#conda install pytables #to save df as hdf5
#conda install -c anaconda tensorflow-gpu
#os.environ['PATH'] = "path-to-openslide-bin" + ";" + os.environ['PATH']
#conda install pytorch==1.7.1 torchvision==0.8.2 torchaudio==0.7.2 cudatoolkit=11.0 -c pytorch

